refactor(pi-codex-conversion): own responses history normalization
