+++ packages/pi-codex-conversion/src/providers/AGENTS.md
@@ -2,5 +2,6 @@
 - Code Mode `exec` uses Pi grammar constrained sampling end-to-end; keep normal mode function-only for parallel tool calls and never restore request-payload contract surgery.
 - Cross-provider replay keeps stable type-correct item IDs (`ctc_*` custom, `fc_*` function) for model switching and cache continuity.
 - `openai-responses/tool-history.ts` is the final pairing boundary after message, replay, and payload rewrites; do not bypass it before transport.
+- `openai-responses/message-history.ts` owns model-switch cleanup, image downgrades, tool-call ID pairing, and synthetic aborted results before `shared.ts` serializes Responses wire items.
 - Configured Code Mode proxies own idempotent Responses Lite relocation/header; never depend on `onPayload`.
 - Configured-provider compaction uses its registered plan-aware stream; Code Mode compaction/replay carries the owned `exec` grammar contract.
