+++ packages/pi-codex-conversion/src/providers/openai-responses/shared.ts
@@ -12,6 +12,7 @@ import {
 } from "../constrained-sampling.js";
 import { parseTextSignature, shortHash } from "./signatures.ts";
 import { normalizeResponsesToolHistory } from "./tool-history.ts";
+import { normalizeResponsesMessageHistory } from "./message-history.ts";
 import { encryptedWebRunOutputFromDetails, imageDetailForResponses, isImageGenerationCallBlock, isWebSearchCallBlock, sanitizeImageGenerationCallItem, sanitizeWebSearchCallItem, type ImageDetail, type ImageGenerationCallBlock, type WebSearchCallBlock } from "./native-items.ts";
 
 type Message = Context["messages"][number];
@@ -85,150 +86,6 @@ function parseResponsesThinkingSignature(signature: string): ResponseInput[numbe
 	}
 }
 
-const NON_VISION_USER_IMAGE_PLACEHOLDER = "(image omitted: model does not support images)";
-const NON_VISION_TOOL_IMAGE_PLACEHOLDER = "(tool image omitted: model does not support images)";
-
-function replaceImagesWithPlaceholder(
-	content: Extract<Message, { role: "user" }> extends { content: infer T } ? Exclude<T, string> : never,
-	placeholder: string,
-) {
-	const result: Array<{ type: "text"; text: string } | { type: "image"; data: string; mimeType: string }> = [];
-	let previousWasPlaceholder = false;
-	for (const block of content) {
-		if (block.type === "image") {
-			if (!previousWasPlaceholder) {
-				result.push({ type: "text", text: placeholder });
-			}
-			previousWasPlaceholder = true;
-			continue;
-		}
-		result.push(block);
-		previousWasPlaceholder = block.text === placeholder;
-	}
-	return result;
-}
-
-function downgradeUnsupportedImages(messages: Context["messages"], model: Model<Api>): Context["messages"] {
-	if (model.input.includes("image")) return messages;
-	return messages.map((msg) => {
-		if (msg.role === "user" && Array.isArray(msg.content)) {
-			return { ...msg, content: replaceImagesWithPlaceholder(msg.content, NON_VISION_USER_IMAGE_PLACEHOLDER) };
-		}
-		if (msg.role === "toolResult") {
-			return { ...msg, content: replaceImagesWithPlaceholder(msg.content, NON_VISION_TOOL_IMAGE_PLACEHOLDER) };
-		}
-		return msg;
-	});
-}
-
-function transformMessages(
-	messages: Context["messages"],
-	model: Model<Api>,
-	normalizeToolCallId?: (id: string, targetModel: Model<Api>, source: Extract<Message, { role: "assistant" }>) => string,
-): Context["messages"] {
-	const toolCallIdMap = new Map<string, string>();
-	const imageAwareMessages = downgradeUnsupportedImages(messages, model);
-	const transformed = imageAwareMessages.map((msg) => {
-		if (msg.role === "user") return msg;
-		if (msg.role === "toolResult") {
-			const normalizedId = toolCallIdMap.get(msg.toolCallId);
-			return normalizedId && normalizedId !== msg.toolCallId ? { ...msg, toolCallId: normalizedId } : msg;
-		}
-		if (msg.role === "assistant") {
-			const assistantMsg = msg;
-			const isSameModel =
-				assistantMsg.provider === model.provider && assistantMsg.api === model.api && assistantMsg.model === model.id;
-			const transformedContent = (assistantMsg.content as InternalAssistantContent[]).flatMap((block) => {
-				if (isImageGenerationCallBlock(block)) return block;
-				if (isWebSearchCallBlock(block)) return block;
-				if (block.type === "thinking") {
-					if (block.redacted) return isSameModel ? block : [];
-					if (isSameModel && block.thinkingSignature) return block;
-					if (!block.thinking || block.thinking.trim() === "") return [];
-					return isSameModel ? block : { type: "text" as const, text: block.thinking };
-				}
-				if (block.type === "text") return isSameModel ? block : { type: "text" as const, text: block.text };
-				if (block.type === "toolCall") {
-					let normalizedToolCall = block;
-					if (!isSameModel && block.thoughtSignature) {
-						normalizedToolCall = { ...block };
-						delete normalizedToolCall.thoughtSignature;
-					}
-					if (!isSameModel && normalizeToolCallId) {
-						const normalizedId = normalizeToolCallId(block.id, model, assistantMsg);
-						if (normalizedId !== block.id) {
-							toolCallIdMap.set(block.id, normalizedId);
-							normalizedToolCall = { ...normalizedToolCall, id: normalizedId };
-						}
-					}
-					return normalizedToolCall;
-				}
-				return block;
-			});
-			return { ...assistantMsg, content: transformedContent as Extract<Message, { role: "assistant" }>["content"] };
-		}
-		return msg;
-	});
-
-	const result: Context["messages"] = [];
-	let pendingToolCalls: Array<Extract<Extract<Message, { role: "assistant" }>["content"][number], { type: "toolCall" }>> = [];
-	let existingToolResultIds = new Set<string>();
-
-	const insertSyntheticToolResults = () => {
-		if (pendingToolCalls.length === 0) return;
-		for (const toolCall of pendingToolCalls) {
-			if (!existingToolResultIds.has(toolCall.id)) {
-				result.push({
-					role: "toolResult",
-					toolCallId: toolCall.id,
-					toolName: toolCall.name,
-					content: [{ type: "text", text: "aborted" }],
-					isError: true,
-					timestamp: Date.now(),
-				});
-				existingToolResultIds.add(toolCall.id);
-			}
-		}
-		pendingToolCalls = [];
-		existingToolResultIds = new Set();
-	};
-
-	for (const msg of transformed) {
-		if (msg.role === "assistant") {
-			insertSyntheticToolResults();
-			if (msg.stopReason === "error" || msg.stopReason === "aborted") continue;
-			const toolCalls = msg.content.filter((block) => block.type === "toolCall");
-			if (toolCalls.length > 0) {
-				const seen = new Set<string>();
-				pendingToolCalls = toolCalls.filter((toolCall) => {
-					if (seen.has(toolCall.id)) return false;
-					seen.add(toolCall.id);
-					return true;
-				});
-				existingToolResultIds = new Set();
-			}
-			result.push(msg);
-			continue;
-		}
-		if (msg.role === "toolResult") {
-			if (!pendingToolCalls.some((toolCall) => toolCall.id === msg.toolCallId) || existingToolResultIds.has(msg.toolCallId)) continue;
-			existingToolResultIds.add(msg.toolCallId);
-			result.push(msg);
-			continue;
-		}
-		if (msg.role === "user") {
-			insertSyntheticToolResults();
-			result.push(msg);
-			continue;
-		}
-		result.push(msg);
-	}
-
-	insertSyntheticToolResults();
-
-	return result;
-}
-
 export function convertResponsesMessages<TApi extends Api>(
 	model: Model<TApi>,
 	context: Context,
@@ -257,7 +114,7 @@ export function convertResponsesMessages<TApi extends Api>(
 		return `${normalizedCallId}|${normalizedItemId}`;
 	};
 
-	const transformedMessages = transformMessages(context.messages, model as Model<Api>, normalizeToolCallId as never);
+	const transformedMessages = normalizeResponsesMessageHistory(context.messages, model as Model<Api>, normalizeToolCallId as never);
 	const includeSystemPrompt = options?.includeSystemPrompt ?? true;
 	if (includeSystemPrompt && context.systemPrompt) {
 		messages.push({ role: model.reasoning ? "developer" : "system", content: sanitizeSurrogates(context.systemPrompt) });
